Router refactor
